experimental/spi/note_assistant_v2/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic[email]
requests
python-dotenv
grpcio
openai
anthropic
google-generativeai
google_auth_oauthlib
shotgun_api3
httpx
websockets
PyYAML>=6.0
